Bug 1993309 - Use test-defined themes instead of built-in themes in webextension action tests. r=tobyp
Thunderbird is doing the right thing, but the tests are using built-in themes which behave slightly
differently from custom themes. The built-in light theme is *supposed* to use the default icon
rather than the dark one, at least that's what Firefox does.

So it's better if the tests define custom themes for checking that the right icon is selected.

Differential Revision: https://phabricator.services.mozilla.com/D268238