Bug 1995024 - Part 2: Add stand-alone NTLM support to ews_xpcom. r=jtracey
As demonstrated in the part 3 commit on this stack, the NTLM authentication flow
might need to be called from separate places, so I thought it best to write most
of the relevant logic in a separate, stand-alone function, which uses Firefox's
NTLM implementation (via `nsIAuthModule`).

See the documentation of the `ntlm::authenticate` function for a rough
documentation of how the NTLM authentication flow is expected to work over HTTP.

This commit also fixes moz_http's handling of multiple headers with the same
name in responses, which is required here since the `WWW-Authenticate` header
might appear multiple times in a response, and one of these occurrences might
contain an NTLM challenge we'll need to authenticate.

Differential Revision: https://phabricator.services.mozilla.com/D270712